+++ register.js
@@ -1,4 +1,0 @@
-'use strict';
-
-const babelSettings = require('./babel-settings.json');
-require('babel-register')(babelSettings);