+++ test/decode.spec.js
@@ -1,6 +1,6 @@
 import should from 'should';
 import lib from '../src/index';
-import constants from './constants';
+import constants from './util/constants';
 
 describe('Decode', () => {
 
@@ -12,8 +12,8 @@ describe('Decode', () => {
 
       (decoded.width).should.be.eql(1052);
       (decoded.height).should.be.eql(1052);
-      (decoded.data).should.be.instanceOf(Uint8Array);
-      done();
+      (decoded.data.buffer).should.be.instanceOf(ArrayBuffer);
+      done(err);
     });
   });
 
@@ -24,8 +24,8 @@ describe('Decode', () => {
       should.exist(decoded);
       (decoded.width).should.be.eql(1052);
       (decoded.height).should.be.eql(1052);
-      (decoded.data).should.be.instanceOf(Uint8Array);
-      done();
+      (decoded.data.buffer).should.be.instanceOf(ArrayBuffer);
+      done(err);
     });
   });
 
@@ -36,8 +36,8 @@ describe('Decode', () => {
       should.exist(decoded);
       (decoded.width).should.be.eql(1052);
       (decoded.height).should.be.eql(1052);
-      (decoded.data).should.be.instanceOf(Uint8Array);
-      done();
+      (decoded.data.buffer).should.be.instanceOf(ArrayBuffer);
+      done(err);
     })
   });
 
@@ -48,8 +48,54 @@ describe('Decode', () => {
       should.exist(decoded);
       (decoded.width).should.be.eql(1052);
       (decoded.height).should.be.eql(1052);
-      (decoded.data).should.be.instanceOf(Uint8Array);
-      done();
+      (decoded.data.buffer).should.be.instanceOf(ArrayBuffer);
+      done(err);
+    });
+  });
+
+  it('fail to process ' + constants.nameAC, (done) => {
+    const jpegData = constants.bufAC;
+    lib.decode(jpegData, (err, decoded) => {
+      should.exist(err);
+      should.not.exist(decoded);
+      err.should.be.an.instanceOf(Error);
+      done(decoded);
+    });
+  });
+
+  it('can be used to process ' + constants.nameP, (done) => {
+    const jpegData = constants.bufP;
+    lib.decode(jpegData, (err, decoded) => {
+      should.not.exist(err);
+      should.exist(decoded);
+      (decoded.width).should.be.eql(1052);
+      (decoded.height).should.be.eql(1052);
+      (decoded.data.buffer).should.be.instanceOf(ArrayBuffer);
+      done(err);
+    });
+  });
+
+  it('can be used to process ' + constants.nameDCTF, (done) => {
+    const jpegData = constants.bufDCTF;
+    lib.decode(jpegData, (err, decoded) => {
+      should.not.exist(err);
+      should.exist(decoded);
+      (decoded.width).should.be.eql(1052);
+      (decoded.height).should.be.eql(1052);
+      (decoded.data.buffer).should.be.instanceOf(ArrayBuffer);
+      done(err);
+    });
+  });
+
+  it('can be used to process ' + constants.nameCP, (done) => {
+    const jpegData = constants.bufCP;
+    lib.decode(jpegData, (err, decoded) => {
+      should.not.exist(err);
+      should.exist(decoded);
+      (decoded.width).should.be.eql(1052);
+      (decoded.height).should.be.eql(1052);
+      (decoded.data.buffer).should.be.instanceOf(ArrayBuffer);
+      done(err);
     });
   });
 
@@ -60,8 +106,8 @@ describe('Decode', () => {
       should.exist(decoded);
       (decoded.width).should.be.eql(1052);
       (decoded.height).should.be.eql(1052);
-      (decoded.data).should.be.instanceOf(Uint8Array);
-      done();
+      (decoded.data.buffer).should.be.instanceOf(ArrayBuffer);
+      done(err);
     });
   });
 
@@ -71,7 +117,7 @@ describe('Decode', () => {
       should.exist(err);
       should.not.exist(decoded);
       err.should.be.an.instanceOf(Error);
-      done();
+      done(decoded);
     });
   });