+++ test/magic.spec.js
@@ -1,11 +1,11 @@
 import should from 'should';
 import lib from '../src/index';
-import constants from './constants';
+import constants from './util/constants';
 
 describe('Magic number', () => {
 
   it('can be detected for a JPEG Buffer', (done) => {
-    const buf = new Buffer([0xFF, 0xD8, 0xFF]);
+    const buf = Buffer.from([0xFF, 0xD8, 0xFF]);
 
     lib.magic(buf, (err, result) => {
       should.not.exist(err);
@@ -14,7 +14,7 @@ describe('Magic number', () => {
       result.mimeType.should.be.equal('image/jpeg');
       result.extension.should.be.equal('jpg');
 
-      done();
+      done(err);
     });
   });
 
@@ -32,7 +32,7 @@ describe('Magic number', () => {
       result.mimeType.should.be.equal('image/jpeg');
       result.extension.should.be.equal('jpg');
 
-      done();
+      done(err);
     });
   });
 
@@ -50,7 +50,25 @@ describe('Magic number', () => {
       result.mimeType.should.be.equal('image/jpeg');
       result.extension.should.be.equal('jpg');
 
-      done();
+      done(err);
+    });
+  });
+
+  it('can be detected for a JPEG Uint8ClampedArray', (done) => {
+    const buf = new ArrayBuffer(3);
+    const view = new Uint8ClampedArray(buf);
+    view[0] = 0xFF;
+    view[1] = 0xD8;
+    view[2] = 0xFF;
+
+    lib.magic(view, (err, result) => {
+      should.not.exist(err);
+      should.exist(result);
+
+      result.mimeType.should.be.equal('image/jpeg');
+      result.extension.should.be.equal('jpg');
+
+      done(err);
     });
   });
 
@@ -64,7 +82,7 @@ describe('Magic number', () => {
       result.mimeType.should.be.equal('image/jpeg');
       result.extension.should.be.equal('jpg');
 
-      done();
+      done(err);
     });
   });
 
@@ -78,7 +96,7 @@ describe('Magic number', () => {
       result.mimeType.should.be.equal('image/png');
       result.extension.should.be.equal('png');
 
-      done();
+      done(err);
     });
   });
 
@@ -89,7 +107,7 @@ describe('Magic number', () => {
       should.exist(err);
       should.not.exist(result);
       err.should.be.an.instanceOf(Error);
-      done();
+      done(result);
     });
   });